Do not use system() call

+++ dynamicc.c
@@ -3,6 +3,7 @@
 #include <unistd.h>
 #include <dlfcn.h>
 #include <errno.h>
+#include <sys/wait.h>
 #include "dynamicc.h"
 
 static bool dc_get_filename(dynamic_c_t *target)
@@ -21,7 +22,7 @@ dynamic_c_t dc_init()
 {
     dynamic_c_t dc = {};
     if (!dc_get_filename(&dc)) {
-        fprintf(stderr, "Cannot get a temp. file!");
+        fprintf(stderr, "Cannot create a temp. C-code file!");
         exit(1);
     }
     else {
@@ -29,7 +30,7 @@ dynamic_c_t dc_init()
     }
 }
 
-void *dc_finish(dynamic_c_t *dc)
+void *dc_finish(dynamic_c_t *dc, const char **cc_args)
 {
     fflush(dc->f);
     fclose(dc->f);
@@ -41,18 +42,21 @@ void *dc_finish(dynamic_c_t *dc)
         return NULL;
     }
 
+    int pid = fork();
+    if (pid == 0) {
+        printf("%s\n", dc->_filename);
+        char *args[] = {"cc", dc->_filename, "-o", so_name, "-Ofast", "-Wall", "-Wextra", "-fPIC", "-shared", NULL};
+        execvp("cc", args);
+        exit(1);
+    }
+    else {
+        int status;
+        waitpid(pid, &status, 0);
+    }
 
-    char command[256];
-    // TODO: Escaping etc.
-    snprintf(command, sizeof(command), "gcc %s -o %s -Ofast -Wall -Werror -Wextra -fPIC -shared", dc->_filename, so_name);
-
-    int err = system(command);
     unlink(dc->_filename);
-    printf("err %d\n", err);
 
-    printf("created %s\n", so_name);
     void *so_handle = dlopen(so_name, RTLD_LAZY);
     unlink(so_name);
-    printf("handle: %p\n", so_handle);
     return so_handle;
 }
